+++ modules/channel-messenger/src/backend/typings.ts
@@ -1,0 +1,6 @@
+import { ChannelContext } from 'common/channel'
+import { MessengerClient } from './messenger'
+
+export type MessengerContext = ChannelContext<MessengerClient> & {
+  messages: any[]
+}