+++ modules/channel-web/src/views/lite/components/messages/renderer/index.tsx
@@ -1,5 +1,6 @@
 export { QuickReplies } from './QuickReplies'
 export { FileMessage } from './FileMessage'
+export { VoiceMessage } from './VoiceMessage'
 export { FileInput } from './FileInput'
 export { Carousel } from './Carousel'
 export { LoginPrompt } from './LoginPrompt'