+++ modules/nlu/src/backend/election/index.ts
@@ -1,10 +1,9 @@
 import * as sdk from 'botpress/sdk'
-import { Config as CoreConfig } from 'common/nlu/engine'
 import { Config as ModConfig } from '../../config'
 import legacyElectionPipeline from './legacy-election'
 import naturalElectionPipeline from './natural-election'
 
-export function election(input: sdk.IO.EventUnderstanding, config: CoreConfig | ModConfig) {
+export function election(input: sdk.IO.EventUnderstanding, config: ModConfig) {
   if (config.legacyElection) {
     return legacyElectionPipeline(input)
   }